moved entity detail page to modules
